% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{aim_add_inventory3d}
\alias{aim_add_inventory3d}
\title{View inventory in 3D}
\usage{
aim_add_inventory3d(x, inventory)
}
\arguments{
\item{x}{A numeric vector for translation offsets. Like in the lidR package.}

\item{inventory}{created by \link{aim_inventory}}
}
\description{
View inventory in 3D
}
