% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matching.R
\name{aim_disentangle_ids}
\alias{aim_disentangle_ids}
\title{Disentangle Tree IDs Between Point Cloud Data and Inventory}
\usage{
aim_disentangle_ids(las, inventory)
}
\arguments{
\item{las}{A \code{LAS} object containing point cloud data with a \code{treeID} attribute.}

\item{inventory}{An \code{sf} object containing forest inventory data with an \code{ID_Arbre} column.}
}
\value{
A \code{LAS} object with updated tree IDs, or the original \code{LAS} object
if no matching IDs are found.
}
\description{
Identifies overlapping tree IDs between a point cloud object and a forest
inventory dataset, remapping conflicting tree IDs.
}
\details{
If a tree has an ID (e.g., 54) in the point cloud and the inventory also
contains an ID = 54, they typically correspond to two different trees (the
probability that they represent the same tree is extremely low). In some
circumstances, this creates entangled IDs. To avoid complex downstream errors,
users can first apply \code{aim_disentangle_ids()}. The function remaps the
IDs in the point cloud such that there is no ID overlap between the point
cloud and the inventory.
}
\seealso{
\code{\link[=aim_inventory]{aim_inventory()}}
}
