% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matching.R
\name{aim_matching_table}
\alias{aim_matching_table}
\title{Create a Tree Matching Table}
\usage{
aim_matching_table(las, inventory)
}
\arguments{
\item{las}{A `LAS` object instance-segmented with the arbor package.}

\item{inventory}{A sf data frame containing the field inventory, including
`ID_Arbre` and `POM` columns. Use \link{aim_inventory}}
}
\value{
A data frame containing the matching results for each inventory tree.
}
\description{
Matches LiDAR-detected trees with trees from a field inventory.
}
