% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matching.R
\name{aim_reassign_ids}
\alias{aim_reassign_ids}
\title{Reassign Tree IDs Using a Matching Table}
\usage{
aim_reassign_ids(las, matching_table)
}
\arguments{
\item{las}{A `LAS` object instance-segmented with the arbor package.}

\item{matching_table}{A data frame containing treeID and targetID
columns defining the ID reassignment.}
}
\value{
A vector with tree IDs reassigned according to matching_table.
}
\description{
Reassigns tree IDs in a vector according to a matching table.
}
