% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matching.R
\name{aim_resolve_multi_matching}
\alias{aim_resolve_multi_matching}
\title{Re-segment Tree IDs Using a Matching Table and an Inventory}
\usage{
aim_resolve_multi_matching(las, matching_table, inventory)
}
\arguments{
\item{las}{A `LAS` object instance-segmented with the arbor package.}

\item{matching_table}{A data frame containing treeID and targetID
columns defining the ID reassignment.}

\item{inventory}{An sf data.frame produced by \link{aim_inventory}.}
}
\value{
A vector with tree IDs reassigned according to matching_table.
}
\description{
Re-segment Tree IDs Using a Matching Table and an Inventory.
}
